+++ kea/hdl/signal_handling/__init__.py
@@ -1,7 +1,8 @@
 from .asynchronous import (
     signal_assigner, constant_assigner, combined_signal_assigner,
     signal_slicer, integer_constant_signal, de_concatenator,
-    DeConcatenatorOutputInterface)
+    DeConcatenatorOutputInterface, sig_assigner, signed_sig_assigner)
 from .synchronous import (
     synchronous_signal_assigner, synchronous_signal_slicer,
-    synchronous_saturating_rounding_slicer)
+    synchronous_saturating_rounding_slicer, sync_left_shift,
+    sync_sig_assigner, sync_sig_assigner_with_reset)