+++ kea/hdl/signal_handling/asynchronous/__init__.py
@@ -5,3 +5,5 @@ from .de_concatenator_block import (
 from ._integer_constant_signal import integer_constant_signal
 from ._signal_assigner import signal_assigner
 from ._signal_slicer import signal_slicer
+from .sig_assigner_block import sig_assigner
+from .signed_sig_assigner_block import signed_sig_assigner