+++ kea/hdl/signal_handling/synchronous/sync_sig_assigner_block/__init__.py
@@ -1,0 +1,1 @@
+from ._sync_sig_assigner import sync_sig_assigner