+++ CHANGELOG.md
@@ -3,6 +3,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.11.0 - 2026-03-17
+
+### Changed
+
+- Renamed `kea/hdl/axi/axi_stream_utils/axis_utils.py` to `kea/hdl/axi/axi_stream_utils/axis_interface_checks.py`.
+- Removed some unused imports.
+- Moved the AXI imports in `/kea/testing/myhdl/cosimulation.py` to solve a circular dependency.
+
+### Added
+
+- Added a asynchronous `sig_assigner` and `signed_sig_assigner`.
+- Added a synchronous `sync_sig_assigner`, `sync_left_shift` and `sync_sig_assigner_with_reset`.
+- Added a `check_bool_or_intbv_signal` to the interface checks.
+
 ## 0.10.0 - 2026-02-27
 
 ### Changed